unify extensions between consumer and producer
